# api/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Access",
    "version": "Not Defined"
  },
  "paths": {
...
    "/api/requests": {
      "get": {
        "tags": [
          "access-requests"
        ],
        "summary": "Access Requests",
        "operationId": "access_requests_api_requests_get",
        "parameters": [
          {
            "name": "q_args",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SearchAccessRequestQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TypeVar_Customized_AccessRequestSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "access-requests"
        ],
        "summary": "Access Requests Create",
        "operationId": "access_requests_create_api_requests_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccessRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRequestSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/requests/{access_request_id}": {
      "get": {
        "tags": [
          "access-requests"
        ],
        "summary": "Access Request By Id",
        "operationId": "access_request_by_id_api_requests__access_request_id__get",
        "parameters": [
          {
            "name": "access_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Access Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRequestDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "access-requests"
        ],
        "summary": "Access Request By Id Put",
        "operationId": "access_request_by_id_put_api_requests__access_request_id__put",
        "parameters": [
          {
            "name": "access_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Access Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAccessRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRequestSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessRequestDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "requester_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester User Id"
          },
          "requested_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Group Id"
          },
          "request_ownership": {
            "type": "boolean",
            "title": "Request Ownership",
            "default": false
          },
          "request_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Reason",
            "default": ""
          },
          "request_ending_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Ending At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "resolver_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolver User Id"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason",
            "default": ""
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "approval_ending_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Ending At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "requester": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_requester": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "requested_group": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/_AccessRequestOktaGroupRef"
                  },
                  {
                    "$ref": "#/components/schemas/_AccessRequestRoleGroupRef"
                  },
                  {
                    "$ref": "#/components/schemas/_AccessRequestAppGroupRef"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "app_group": "#/components/schemas/_AccessRequestAppGroupRef",
                    "okta_group": "#/components/schemas/_AccessRequestOktaGroupRef",
                    "role_group": "#/components/schemas/_AccessRequestRoleGroupRef"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Group"
          },
          "active_requested_group": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OktaGroupRef"
                  },
                  {
                    "$ref": "#/components/schemas/RoleGroupRef"
                  },
                  {
                    "$ref": "#/components/schemas/AppGroupRef"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "app_group": "#/components/schemas/AppGroupRef",
                    "okta_group": "#/components/schemas/OktaGroupRef",
                    "role_group": "#/components/schemas/RoleGroupRef"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Requested Group"
          },
          "resolver": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_resolver": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AccessRequestDetail",
        "description": "Response for `GET /api/requests/{id}` only.\n\nFlask `AccessRequestResource.get()` `only=` retains a richer\n`requested_group` projection (group tags + role-association mappings on\nrole groups, plus `app`/`is_owner` on app groups). We split the detail\nshape from the list/POST/PUT shape (`AccessRequestSummary`) because\nSQLAlchemy `lazy=\"raise_on_sql\"` won't let one Pydantic model\nconditionally include the relationships at runtime \u2014 the eager-loads\ndiffer per endpoint."
      },
      "AccessRequestSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "requester_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester User Id"
          },
          "requested_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Group Id"
          },
          "request_ownership": {
            "type": "boolean",
            "title": "Request Ownership",
            "default": false
          },
          "request_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Reason",
            "default": ""
          },
          "request_ending_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Ending At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "resolver_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolver User Id"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason",
            "default": ""
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "approval_ending_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Ending At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "requester": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_requester": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "requested_group": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OktaGroupRef"
                  },
                  {
                    "$ref": "#/components/schemas/RoleGroupRef"
                  },
                  {
                    "$ref": "#/components/schemas/AppGroupRef"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "app_group": "#/components/schemas/AppGroupRef",
                    "okta_group": "#/components/schemas/OktaGroupRef",
                    "role_group": "#/components/schemas/RoleGroupRef"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Group"
          },
          "active_requested_group": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OktaGroupRef"
                  },
                  {
                    "$ref": "#/components/schemas/RoleGroupRef"
                  },
                  {
                    "$ref": "#/components/schemas/AppGroupRef"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "app_group": "#/components/schemas/AppGroupRef",
                    "okta_group": "#/components/schemas/OktaGroupRef",
                    "role_group": "#/components/schemas/RoleGroupRef"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Requested Group"
          },
          "resolver": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_resolver": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AccessRequestSummary",
        "description": "Response for `GET /api/requests`, `POST /api/requests`, and\n`PUT /api/requests/{id}`. The `requested_group` is the lightweight\n`GroupRef` \u2014 no group tags or role-association mappings, which only\nthe detail page renders. Absent optional fields are omitted from the\nresponse body (the ExcludeNoneAPIRoute default), so clients should\ntreat missing keys as None."
      },
...
      "CreateAccessRequestBody": {
        "properties": {
          "group_id": {
            "type": "string",
            "title": "Group Id"
          },
          "group_owner": {
            "type": "boolean",
            "title": "Group Owner",
            "default": false
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "default": ""
          },
          "ending_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ending At"
          }
        },
        "type": "object",
        "required": [
          "group_id"
        ],
        "title": "CreateAccessRequestBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_TypeVar_Customized_AccessRequestSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccessRequestSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[TypeVar]Customized[AccessRequestSummary]"
      },
...
      "SearchAccessRequestQuery": {
        "properties": {
          "q": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "requester_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester User Id"
          },
          "assignee_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee User Id"
          },
          "requested_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Group Id"
          },
          "resolver_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolver User Id"
          }
        },
        "type": "object",
        "title": "SearchAccessRequestQuery"
      },
...
    }
  }
}